Publication commit: init, incl license & readme
This commit initializes the repository with the code at the time of
publication acceptance by Journal of Ecology. It includes a reorganization
of the files for easy access under a single repository (BucketsMarkov was
previously separate), and a renaming of the base directory. Some comments
have been added, old files and results removed from version control, and
both a README.md and license.txt have been added for publication on
Github. Actual journal figures and simulation results have been ommited
except in the case of the BucketsMarkov TBA parameter sweeps. All rainfall
data is included, and results should be fully reproducable using only the
tracked files.